+++ install.sh
@@ -5,8 +5,13 @@ DST_DIR="${HOME}"
 printf "${LABEL} home directory: ${DST_DIR}\n"
 
 function fsync {
-  printf "${LABEL} ${1} ${2}\n"
-	rsync $1 $2
+    # check if the file exists
+    if [ ! -f $1 ]; then
+        return 0
+    fi
+
+    printf "${LABEL} ${1} ${2}\n"
+    rsync $1 $2
 }
 
 fsync emacs/.emacs "${DST_DIR}"
@@ -14,25 +19,34 @@ fsync vim/.vimrc "${DST_DIR}"
 fsync bash/.bashrc "${DST_DIR}"
 fsync bash/.bash_profile "${DST_DIR}"
 fsync bash/.inputrc "${DST_DIR}"
+
+SHELLFILES_DIR="${DST_DIR}/.shell_files"
+mkdir -p "${SHELLFILES_DIR}"
+fsync bash/alias_command.bash "${SHELLFILES_DIR}"
+
 fsync .tmux.conf "${DST_DIR}"
 fsync .pylintrc "${DST_DIR}"
 
+##################################################
+# nvim config
 CONFIG_DIR="${DST_DIR}/.config"
-if [ ! -d "${CONFIG_DIR}" ]; then
-  mkdir "${CONFIG_DIR}"
-fi
+mkdir -p "${CONFIG_DIR}"
 
 NVIM_DIR="${CONFIG_DIR}/nvim"
-if [ ! -d "${NVIM_DIR}" ]; then
-  mkdir "${NVIM_DIR}"
-fi
+mkdir -p "${NVIM_DIR}"
+
 fsync vim/init.vim "${NVIM_DIR}"
 
-JUPYTERNB_DIR="${DST_DIR}/.jupyter/nbconfig/"
-if [ ! -d "${JUPYTERNB_DIR}" ]; then
-  mkdir "${JUPYTERNB_DIR}"
-fi 
+##################################################
+# jupyter notebook
+JUPYTERNB_DIR="${DST_DIR}/.jupyter"
+mkdir -p "${JUPYTERNB_DIR}"
+
+JUPYTERNBCONFIG_DIR="${DST_DIR}/.jupyter/nbconfig/"
+mkdir -p "${JUPYTERNBCONFIG_DIR}"
+
 fsync jupyter/notebook.json "${JUPYTERNB_DIR}"
+##################################################
 
 printf "${LABEL} done\n"
